# openapi/api.yaml
...
openapi: 3.0.0
info:
  version: 0.0.1
  title: Greenlake platform API for returning metadata and configuration for colo deployments
  description: >-
    NOTE: These are proposed APIs for gathering metadata information for PCE service.
  license:
    name: HPE End User License Agreement
    url: https://www.hpe.com/us/en/software/licensing.html
servers:
  - url: https://us-east1.api.privatecloud-ent.greenlake.hpe.com
security:
  - bearerAuth: []
paths:
  /api/v1alpha1/providers/:
    post:
      tags:
        - v1alpha1
      operationId: registerProvider
      summary: Configure a Colo provider
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderRegistrationInput"
      responses:
        200:
          $ref: "#/components/responses/RegistrationResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        409:
          $ref: "#/components/responses/Conflict"
        500:
          $ref: "#/components/responses/InternalServerError" 
    get:
      tags:
        - v1alpha1
      operationId: ListProviders
      summary: Get list of supported provider and regions
      parameters:
        - in: query
          name: next
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 100
            maximum: 200
      responses:
        200:
          $ref: "#/components/responses/ProviderListResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/InternalServerError"
  /api/v1alpha1/locations/:
    get:
      tags:
        - v1alpha1
      operationId: ListLocations
      summary: Get list of location for a specific provider
      parameters:
        - in: query
          name: next
          required: false
          schema:
           type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 100
            maximum: 200
        - in: query
          name: providerId
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: "#/components/responses/LocationListResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/InternalServerError"
  /api/v1alpha1/locations/{id}/devices/:
    get:
      tags:
        - v1alpha1
      operationId: getManagedDevicesById
      summary: Get managed devices in a location by Id
      parameters:
        - in: path
          name: id
          description: Id of a specific location
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          $ref: "#/components/responses/DeviceListResponse"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
  /api/v1alpha1/locations/{id}/racks/:
    get:
      tags:
        - v1alpha1
      operationId: getManagedRacksById
      summary: Get managed Racks by a location Id
      parameters:
        - in: path
          name: id
          description: Id of a specific location
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          $ref: "#/components/responses/RackListResponse"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"         
  /api/v1alpha1/services/inventory/actions/get-all/:
    get:
      tags:
        - v1alpha1
      operationId: getAllInventory
      summary: Get all inventory
      parameters:
        - in: query
          name: next
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 100
            maximum: 200     
      responses:
        200:
          $ref: "#/components/responses/HeirarchicalInventoryResponse"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"         
  /api/v1alpha1/deployments/:
    get:
      tags:
        - v1alpha1
      operationId: getManagedDeployments
      summary: Get all Deployments filtered by a location Id
      parameters:
        - in: query
          name: id
          description: Id of a specific location instance
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          $ref: "#/components/responses/DeploymentListResponse"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"         
  /api/v1alpha1/deployments/{id}/:
    get:
      tags:
        - v1alpha1
      operationId: getManagedDeploymentById
      summary: Get a Deployment by Id
      parameters:
        - in: path
          name: id
          description: Id of a specific deployment instance
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          $ref: "#/components/responses/DeploymentListResponse"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"         
components:
  securitySchemes:
    bearerAuth:
      name: Authorization
      in: header
      type: apiKey
  responses:
    RegistrationResponse:
      description: colo provider record
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ColoProviderEntry"
      headers:
        Location:
          description: URL of the created Workspace
          schema:
            type: string    
    ProviderListResponse:
      description: List of colo providers registered with HPE
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProviderList"           
    LocationListResponse:
      description: List of registered locations with a provider
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LocationList"           
    DeviceListResponse:
      description: List of registered devices in a location
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeviceList"     
    RackListResponse:
      description: List of registered racks in a location
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RackList"     
    DeploymentListResponse:
      description: List of all Greenlake deployments in a location
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DeploymentList"     
    NotFound:
      description: 404 Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: 400 Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestError"
    Unauthorized:
      description: 401 Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 403 Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 409 Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: 500 Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalServerError"
    HeirarchicalInventoryResponse:
      description: List of all inventory details in a hierarchical form
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HeirarchicalInventory"                 
  schemas:
    ProviderRegistrationInput:
      type: object
      required:
        - providerName
        - providerUrl
        - providerCred
      properties:
        providerName:
          type: string
          example: equinix
        providerUrl:
          type: string
          format: url
          example: api.dcim.io
        providerCred:
          type: object
          properties:
            secretId:
              type: string
            secretKey:
              type: string
    ProviderList:
      type: object
      required:
        - items
        - count
      properties:
        items:
          type: array
          description: list of registered provider
          items:
            $ref: "#/components/schemas/ColoProviderEntry"
        next:
          type: string
          description: >-
            Specifies the pagination cursor for the next page of resources. Will be 'null' if
            there are no further pages
          example: "c8303c9c-bf2f-4bd8-b8e4-205300175694"
        count:
          type: integer
          description: >-
            The number of provider entries returned
          example: 1         
    LocationList:
      type: object
      required:
        - items
        - count
      properties:
        items:
          type: array
          description: list of locations for a given Provider
          items:
            $ref: "#/components/schemas/LocationEntry"
        next:
          type: string
          description: >-
            Specifies the pagination cursor for the next page of resources. Will be 'null' if
            there are no further pages
          example: "c8303c9c-bf2f-4bd8-b8e4-205300175694"
        count:
          type: integer
          description: >-
            The number of drift entries returned
          example: 1         
    DeviceList:
      type: object
      required:
        - items
        - count
      properties:
        items:
          type: array
          description: list of managed workspaces for the given User
          items:
            $ref: "#/components/schemas/DeviceEntry"
        next:
          type: string
          description: >-
            Specifies the pagination cursor for the next page of resources. Will be 'null' if
            there are no further pages
          example: "c8303c9c-bf2f-4bd8-b8e4-205300175694"
        count:
          type: integer
          description: >-
            The number of drift entries returned
          example: 1         
    RackList:
      type: object
      required:
        - items
        - count
      properties:
        items:
          type: array
          description: list of managed workspaces for the given User
          items:
            $ref: "#/components/schemas/RackEntry"
        next:
          type: string
          description: >-
            Specifies the pagination cursor for the next page of resources. Will be 'null' if
            there are no further pages
          example: "c8303c9c-bf2f-4bd8-b8e4-205300175694"
        count:
          type: integer
          description: >-
            The number of drift entries returned
          example: 1         
    DeploymentList:
      type: object
      required:
        - items
        - count
      properties:
        items:
          type: array
          description: list of deployments. Each deployment may be PCE or PCBe
          items:
            $ref: "#/components/schemas/DeploymentEntry"
        next:
          type: string
          description: >-
            Specifies the pagination cursor for the next page of resources. Will be 'null' if
            there are no further pages
          example: "c8303c9c-bf2f-4bd8-b8e4-205300175694"
        count:
          type: integer
          description: >-
            The number of drift entries returned
          example: 1
    ColoProviderEntry:
      type: object
      required:
        - id
        - createdAt
        - credentialsId
      properties:
        displayName:
          type: string
          description: Unique display name of the colo provider
          example: "Equinix"     
        id:
          type: string
          format: uuid
          description: Unique ID of colo provider
          example: "05f0523c-fd03-47fc-981b-9c4333a37b75"
        apiurl:
          type: string
          format: url
          description: Url to access API for the provider
          example: https://us-east-1.api-seller.equinix.com
        providerProperties:
          type: object
          description: Url to access API for the provider
          properties:
            prop1:
              type: string
            prop2:
              type: string
        createdAt:
          type: string
          format: date-time
          description: Time of resource creation in Greenlake
          readOnly: true
        credentialsId:
          type: string
          format: uuid
          description: The ID of credentials to retrieve the secret from vault
          readOnly: true
...
    HeirarchicalInventory:
      type: object
      properties:
        items:
          type: array
          description: list of facility
          items:
            $ref: "#/components/schemas/ColoFacility"
        count:
          type: integer
          description: >-
            The number of region resource entries returned
          example: 5
...
    BadRequestError:
      type: object
      required:
        - httpStatusCode
        - errorCode
        - message
        - debugId
      properties:
        httpStatusCode:
          type: integer
          description: The HTTP equivalent status code
          example: 400
        errorCode:
          type: string
          description: A unique machine-friendly identifier for the error from a global list of enumerated identifier strings
          example: "HPE_ORGS_EXPIRED_TOKEN"
        message:
          type: string
          description: A user-friendly error message
          example: "Authentication error"
        debugId:
          type: string
          description: A unique identifier for the instance of this error
          example: "12312-123123-123123-1231212"
        errorDetails:
          type: array
          description: Additional detailed information about the error
          items:
            type: object
            required:
              - type
              - issues
            properties:
              type:
                type: string
                description: The type of error details
                example: "hpe.greenlake.bad_request"
              issues:
                type: array
                description: Array of bad request issues
                items:
                  $ref: "#/components/schemas/BadRequestIssues"
    BadRequestIssues:
      type: object
      required:
        - source
        - subject
      properties:
        source:
          type: string
          description: The part of the request with an issue
          enum:
            - field
            - header
            - query.parameter
          example: "field"
        subject:
          type: string
          description: The specific issue key
          example: "user.email"
        description:
          type: string
          description: An elaborate description of the issue. This can be used by developers to understand how the failure can be addressed
          example: "Invalid format: email address length must be between 2 and 320 characters"
    InternalServerError:
      type: object
      required:
        - httpStatusCode
        - errorCode
        - message
        - debugId
      properties:
        httpStatusCode:
          type: integer
          description: The HTTP equivalent status code
          example: 400
        errorCode:
          type: string
          description: A unique machine-friendly identifier for the error from a global list of enumerated identifier strings
          example: "HPE_ORGS_EXPIRED_TOKEN"
        message:
          type: string
          description: A user-friendly error message
          example: "Authentication error"
        debugId:
          type: string
          description: A unique identifier for the instance of this error
          example: "12312-123123-123123-1231212"
        errorDetails:
          type: array
          description: Additional detailed information about the error
          items:
            type: object
            required:
              - type
              - retryAfterSeconds
            properties:
              type:
                type: string
                description: The type of error details
                example: "hpe.greenlake.retry_info"
              retryAfterSeconds:
                type: integer
                description: Seconds to wait before retrying
                example: 30
                default: 30
    Error:
      type: object
      required:
        - httpStatusCode
        - errorCode
        - message
        - debugId
      properties:
        httpStatusCode:
          type: integer
          description: The HTTP equivalent status code
          example: 401
        errorCode:
          type: string
          description: A unique machine-friendly identifier for the error from a global list of enumerated identifier strings
          example: "GLCP-0001"
        message:
          type: string
          description: A user-friendly error message
          example: "Authentication error"
        debugId:
          type: string
          description: A unique identifier for the instance of this error. Maybe same as trace Id
          example: "12312-123123-123123-1231212"
        errorDetails:
          type: array
          description: Additional detailed information about the error
          items:
            type: object
            required:
              - type
              - source
              - metadata
            properties:
              type:
                type: string
                description: The type of error details
                example: "hpe.greenlake.metadata"
              source:
                type: string
                description: The source of the error. Typically a registered API group
                example: "hpe.greenlake.iam"
              metadata:
                type: object
                description: Any additional key value pairs that the service defines
                properties:
                  recommendedActions:
                    type: array
                    items:
                      type: string
                      description: Action recommended to the user
                      example: "Contact admin to perform operation"